requirements.txt
# OCR
pytesseract==0.2.6
typing==3.7.4.3
python-multipart==0.0.5

# Tables
numpy==1.22.0
pandas==1.3.5
xlrd-2.0.1

# Para la API de salida
pydantic==1.9.0
asgiref==3.4.1
certifi==2021.5.30
click==7
fastapi==0.70.1
h11==0.12.0
pydantic==1.9.0
starlette==0.16.0
typing-extensions
uvicorn==0.16.0

# Camelot and Excalibur requerimenst
amqp==5.0.6
camelot-py[cv]==0.9.0
chardet==4.0.0
cryptography==3.4.8
et-xmlfile==1.1.0
greenlet==1.1.2
markupsafe==2.0.1
opencv-python==4.5.4.60
openpyxl==3.0.7
pandas==1.3.5
pdfminer.six==20201018
pypdf2==1.26.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sortedcontainers==2.4.0
waitress==2.0.0
wcwidth==0.2.5
